Moved parsers to their own directory and moved logic to serialize pokemon into their own helper 'module'
